create new training databases for both clf and rpn with validation data

+++ tools/db_creation.py
@@ -11,13 +11,13 @@ import numpy as np
 import os
 
 from lib.utils.load_image import load_image
-from tools.data_parser import parse_csv
+from tools.data_parser import get_csv_data
 
 
 def create_training_db(pct_train, db_type="rpn", use_previous=True):
 
     # Load csv
-    data = parse_csv()
+    data = get_csv_data()
 
     db, ids = [], set()
     for filename in os.listdir("annotations"):
@@ -71,10 +71,13 @@ def store_images():
         # Skip other files
         if not filename.endswith('.nii.gz'):
             continue
+        # Check if file already exists
+        im_name = filename.split(".")[0]
+        if os.path.exists(os.path.join(im_dir, "%s.npy" % im_name)):
+            continue
         # Load image
         im = load_image(os.path.join(data_dir, filename))
         # Store image
-        im_name = filename.split(".")[0]
         np.save(os.path.join(im_dir, "%s.npy" % im_name), im)
 
 
@@ -101,6 +104,12 @@ def split_and_save_db(database, ids, pct_train, db_type, use_previous):
     np.save(os.path.join(extraction_dir, "imdb_val_%s_radio.npy" % db_type), imdb_val)
     np.save(os.path.join(extraction_dir, "imdb_test_%s_radio.npy" % db_type), imdb_test)
     np.save(os.path.join(extraction_dir, "ids_%s_radio.npy" % db_type), ids)
+
+    # Get im names to be able to keep the same split
+    train_names = [im["name"].split("/")[-1].split(".")[0] for im in imdb_train]
+    test_names = [im["name"].split("/")[-1].split(".")[0] for im in imdb_test]
+    np.save(os.path.join("database", "train_names.npy"), train_names)
+    np.save(os.path.join("database", "test_names.npy"), test_names)
 
 
 def get_split_idx(database, pct_train, use_previous):
